refactor: extract ihex_decode to decodeHexLine
Only added one test as there is no branching in the code
